Вынес главный экран в отдельный MainFragment
