+++ app/src/main/java/com/geekbrains/potd/MainActivity.kt
@@ -6,19 +6,11 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import com.geekbrains.potd.databinding.ActivityMainBinding
-import com.geekbrains.potd.fragments.earth.EarthFragment
-import com.geekbrains.potd.fragments.mars.MarsFragment
-import com.geekbrains.potd.fragments.system.SystemFragment
 
 class MainActivity : AppCompatActivity() {
     val mainViewModel: MainViewModel by viewModels()
 
     private lateinit var binding: ActivityMainBinding
-    private val navFragments = mapOf(
-        Pair(R.id.bottom_view_system, SystemFragment()),
-        Pair(R.id.bottom_view_earth, EarthFragment()),
-        Pair(R.id.bottom_view_mars, MarsFragment()),
-    )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -26,14 +18,7 @@ class MainActivity : AppCompatActivity() {
         mainViewModel.themeId?.let { setTheme(it) }
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        binding.bottomNavigationView.setOnItemSelectedListener {
-            navFragments[it.itemId]?.let { fragment ->
-                supportFragmentManager.beginTransaction().replace(R.id.container, fragment).commit()
-                true
-            } ?: false
-        }
         binding.topToolbar.setOnMenuItemClickListener { onOptionsItemSelected(it) }
-        binding.bottomNavigationView.selectedItemId = R.id.bottom_view_system
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
