--- app/src/main/res/layout/activity_api_tabitem_system.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.appcompat.widget.AppCompatTextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:drawableStart="@drawable/ic_tabitem_system"
-        android:drawablePadding="@dimen/tab_label_ic_padding"
-        android:gravity="center_vertical"
-        android:text="@string/label_tabitem_system"
-        android:textColor="?attr/colorPrimary"
-        android:textSize="@dimen/tab_label_text_size"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
